set up file upload and download

+++ controllers/admin.js
@@ -1,17 +1,51 @@
+const { validationResult } = require("express-validator");
+// const { Types } = require("mongoose");
 const Product = require("../models/product");
+
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
     path: "/admin/add-product",
     pageTitle: "Add Product",
-    editing: false
+    editing: false,
+    errorMessage: "",
+    oldInput: {},
+    validationErrors: []
   });
 };
 
 exports.postAddproduct = async (req, res, next) => {
-  const { title, imageUrl, price, description } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+  // console.log(image);
+
   try {
+    if (!image) {
+      return res.status(422).render("admin/edit-product", {
+        path: "/admin/add-product",
+        pageTitle: "Add Product",
+        editing: false,
+        errorMessage: "Attached file is not an image",
+        oldInput: { title, price, description },
+        validationErrors: []
+      });
+    }
+
+    const imageUrl = image.path;
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(422).render("admin/edit-product", {
+        path: "/admin/add-product",
+        pageTitle: "Add Product",
+        editing: false,
+        errorMessage: errors.array()[0].msg,
+        oldInput: { title, image, price, description },
+        validationErrors: errors.array()
+      });
+    }
     await new Product({
+      // _id: Types.ObjectId("5e624813e5dea53b04fb1a03"),
       title,
       price,
       description,
@@ -20,7 +54,9 @@ exports.postAddproduct = async (req, res, next) => {
     }).save();
     res.redirect("/admin/products");
   } catch (err) {
-    console.log(err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
   }
 };
 
@@ -40,37 +76,67 @@ exports.getEditProduct = async (req, res, next) => {
     path: "/admin/edit-product",
     pageTitle: "Edit Product",
     editing: editMode,
-    product: product
+    product: product,
+    errorMessage: "",
+    oldInput: {},
+    validationErrors: []
   });
 };
 
 exports.postEditProduct = async (req, res, next) => {
-  const { productId, title, imageUrl, price, description } = req.body;
+  const { productId, title, price, description } = req.body;
+
+  const image = req.file;
   try {
     const product = await Product.findById(productId);
 
     if (product.userId.toString() !== req.user._id.toString()) {
       return res.redirect("/");
     }
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(422).render("admin/edit-product", {
+        path: "/admin/edit-product",
+        pageTitle: "Edit Product",
+        editing: true,
+        product: { title, price, description, _id: productId },
+        errorMessage: errors.array()[0].msg,
+        oldInput: {},
+        validationErrors: errors.array()
+      });
+    }
     product.title = title;
-    product.imageUrl = imageUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     product.price = price;
     product.description = description;
     await product.save();
     res.redirect("/admin/products");
   } catch (err) {
-    console.log(err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
   }
 };
 
 exports.postDeleteProduct = async (req, res, next) => {
   const { productId } = req.body;
   try {
+    const product = await Product.findById(productId);
+    if (!product) {
+      return next(new Error("product not found"));
+    }
+    fileHelper.deleteFile(product.imageUrl);
     await Product.deleteOne({ _id: productId, userId: req.user._id });
     console.log("DELETED PRODUCT");
     res.redirect("/admin/products");
   } catch (err) {
-    console.log(err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
   }
 };
 
@@ -83,6 +149,8 @@ exports.getAdminProducts = async (req, res, next) => {
       pageTitle: "Admin Products"
     });
   } catch (err) {
-    console.log(err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
   }
 };
